Version Control for Game Demos and Projects
Added a version control system for the game Demo and game project locations.
System includes adding a button for each version that shows the game version that it describes.
Cleaned up the Star Wars Cantina Games page.
Added the Project Equinox game page, will need to complete this and link it to the home page.